# yiba-verified/docs/09-API/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Yiba Verified API
  description: |
    Comprehensive API for managing educational institutions, learners, enrolments, submissions, and QCTO workflows.
    
    ## Authentication
    
    ### Development Mode
    - Use `X-DEV-TOKEN` header with your `DEV_API_TOKEN` value
    - Only works in development environment (`NODE_ENV=development`)
    - Example: `X-DEV-TOKEN: your-dev-token-here`
    
    ### Production Mode
    - Use NextAuth session cookies
    - Login via `/auth/login` endpoint
    - Session cookie is automatically included in requests
    
    ## Role-Based Access Control (RBAC)
    
    - **PLATFORM_ADMIN**: Full access to all resources across all institutions
    - **QCTO_USER**: Read-only access to approved submissions and requests
    - **INSTITUTION_ADMIN**: Full access to their institution's resources
    - **INSTITUTION_STAFF**: Limited access to their institution's resources
    - **STUDENT**: Access only to their own data
    
    ## Rate Limiting
    
    - **Auth endpoints**: 5 requests per 15 minutes
    - **Mutations**: 30 requests per minute
    - **Uploads**: 10 requests per minute
    - **Standard API**: 60 requests per minute
    - **Reads**: 120 requests per minute
    
  version: 1.0.0
  contact:
    name: Yiba Verified Support
    email: support@yiba.co.za
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.yiba.co.za/api
    description: Production server
...
security:
  - DevToken: []
  - SessionCookie: []
paths:
...
  /export/learners:
    get:
      tags: [Export]
      summary: Export learners data
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, json]
        - name: institution_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Export file
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Learner'
components:
  securitySchemes:
    DevToken:
      type: apiKey
      in: header
      name: X-DEV-TOKEN
      description: Development token (dev mode only)
    SessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
    Learner:
      type: object
      properties:
        learner_id:
          type: string
          format: uuid
        national_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        birth_date:
          type: string
          format: date
        institution_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...